refactor(composer-app): add util/index.ts barrel and import from util not util/*
