+++ packages/apps/composer-app/src/recovery/boot-client.ts
@@ -7,7 +7,7 @@ import { mountDevtoolsHooks } from '@dxos/client/devtools';
 import { Config, defs } from '@dxos/config';
 import { Runtime } from '@dxos/protocols/proto/dxos/config';
 
-import { setupConfig } from '../util/config';
+import { setupConfig } from '../util';
 
 let bootedClient: Client | undefined;
