+++ packages/apps/composer-app/src/recovery/download-logs.ts
@@ -4,8 +4,7 @@
 
 import { IdbLogStore } from '@dxos/log-store-idb';
 
-import { LOG_STORE_DB_NAME } from '../util/constants';
-import { downloadLogs } from '../util/log-download';
+import { LOG_STORE_DB_NAME, downloadLogs } from '../util';
 
 /** Export NDJSON logs from the IDB log collector and trigger a browser download. */
 export const downloadRecoveryLogs = async (): Promise<{ byteLength: number }> => {
