+++ packages/apps/composer-app/src/workers/dedicated-worker.ts
@@ -7,8 +7,7 @@ import { log } from '@dxos/log';
 import { IdbLogStore } from '@dxos/log-store-idb';
 import { isTauri } from '@dxos/util';
 
-import { initializeObservability } from '../util/config';
-import { LOG_STORE_DB_NAME } from '../util/constants';
+import { LOG_STORE_DB_NAME, initializeObservability } from '../util';
 
 const logStore = new IdbLogStore({ dbName: LOG_STORE_DB_NAME });
 log.addProcessor(logStore.processor);
